Remove obsolete heap scan readahead code.
The main effect of the readahead code was not readahead as such, but to
avoid trashing the buffer cache when doing a sequential scan of a large
table. We just got similar functionality from PostgreSQL 8.3: the "ring
buffer" feature, in commit d526575f. So we don't need the GPDB-specific
code anymore.

This was slightly mismerged: we were reading the same page twice in a row:
once with the new ReadBufferWithStrategy() call and second time with the
old KillAndReadBuffer() call. That is now fixed too.