Use empty target list in "dummy" initplans.
We replace unused init plans with dummy Result plans at end of planning.
Use an empty target list for the dummy Result, instead the target list of
the original Plan we're replacing. An empty target list is obviously better
from a performance point of view, but more importantly, if the original
target list contained an AggRef node, initializing the AggRef expression
at ExecutorStart() will fail, because there is a check that an AggRef
can only appear in an Agg node.

Per Nikhil Kak's report. gpcheckcat contains a query that tripped on this.