Remove inaccurate icg test that injects QueryFinishPending
This test was introduced in commit c36ecd6f6fafa64840aac793590988ecfe329751.
It does not actually test the changes of that commit, instead it was aiming to test some other scenarios.
It seems that the test does not have a deterministic behavior.
We remove it for now, and we will investigate if the scenarios that we were aiming to test are possible and write deterministic tests.