Remove and rewrite logically inverted if else clauses
Normally the compiler can figure out when logical chains have inverted
truths like: if (foo) {} else if (!foo) {}. When there are more variables
involved this doesn't always work and we get compiler warnings for unset
variables which are indeed set. Remove and rewrite offending logic to be
clearer for the compiler and the human reader.