Avoid shifting a negative value
Shifting a negative signed value is undefined behavior in C, instead
replace with the defined constant that was intended to reach with
the bitshift. It's not entirely clear to me *why* setting such a low
value is of interest given that further operations on it can overflow
to being positive but this keeps the current functionality with one
less compiler warning.