Avoid writing past the end of the array in workfile stats
When the workfile caching support was removed the number of columns
in the gp_workfile_mgr_cache_entries view was decreased to twelve.
Fix the array offsets to avoid writing past the end of the array as
they were accidentally left out in that commit.

Also move the assertion for view width to before the use of the
tupledesc's in question as it's more useful there than after we've
accessed what we're asserting.