Remove dead code in partition coalesce functionality
The ALTER TABLE .. COALESCE PARTITION feature is while partially
implemented not supported. Removing all the scaffolding around the
parsing might as well be worthwhile but at least it seems reasonable
to kill the completely dead code in ATPExecPartCoalesce(). As this
was the only external caller of parruleord_open_gap() make the
function static.