Use ExprTreeGeneratorInfo for expression tree generation.
ExprTreeGeneratorInfo object keeps track of metadata information
needed to generate code for expression trees.
Refactor ExprTreeGenerators to always use econtext->ecxt_scantuple to
generate the functions. We also have an extra slot_check block to fall
back if it is called with a different slot - which shouldn't happen for
Agg and Scan nodes.