Fix some amount of offending whitespace and style in atmsort
This is by no means all the offending lines in this file, but they
were the ones I had to read in order to figure out a bug. Remove
the worst offenders such as three-space indentation, commented out
code and and blocks at the wrong level. A larger cleanup is at some
point needed but piece by piece we are getting this code into shape.