Avoid accessing array outside array bounds
The @outarr array may at this point be empty and using subscript -1
to reach the last element will thus throw an uninitialized error.
Test for array contents before accessing any elements to avoid the
error and to exit early since the block by definition isn't of
interest. Observed in the wild for the bfv_partition test suite.