Fix bug in gpdbrestore -S schema level restore with truncate option
This commit fixes a bug when performing a gpdbrestore -S on a schema with
the truncate option. The restore would fail stating this option was unsupported.
This code was previously not reached and was not properly escaped for
special characters. Behave and unit tests are added.

Authors: Chris Hajas and Chumki Roy