Remove support for ignore_headers command in gpdiff
gpdiff had support for ignoring header lines in query output which
we aren't using in any testsuite at all. Since this required regex
checking for each line in the output (it could be turned on for a
section of the output) this feature has a non-trivial cost. Remove
as an optimization for the tests that we do have.