Add EXPLAIN CODEGEN to print generated IR to the client
This will aid tremendously in debugging codegen because the generated IR
will be easily available to be seen. Rest of the functionality of
EXPLAIN is maintained. Also we only perform the compilation step if it
not a EXPLAIN-only query. That is EXPLAIN CODEGEN with perform code
generation only and will not compile the resultant code. EXPLAIN ANALYZE
will do both, and also execute the query.