Assert rather than returning NULL for scan method check
Since no callsite of these functions checks for a NULL return value
and instead deref the returned pointer immediately we need to ensure
that only valid values are ever returned. This also corrects a logic
error in the enum range check which was introduced with the recent
effort to silence compiler warnings.