Make static compilation of ORCA work.
Reorder -lgpopt, -lnaucrates, -lgpdbcost, -lgpos, to make static
compilation of ORCA to work. When linking with dynamic libraries, the order
doesn't matter, but with static libraries, the options need to be given in
dependency order.

This doesn't change the way those libraries are actually linked. Whether
we link with static or shared versions depends on which ones are found
at link time.