[#123259241] Fixes partition filters with array constraints
Allowing array constraints to be derived on expressions exposed a situation
where array constraints propagated to partition filters. Arrays weren't
previously handled by the Expression to DXL translator, they were processed
into disjunctions just prior to translation. The change we implemented
is twofold:

(1) Allow array expressions to be expanded out into OR expressions during DXL
translation.

This avoids any error conditions when array expressions appear in partition
filters because the DXL translator can successfully translate it. Note
that there is another method which could have been used for expansion:
CScalarArrayCmp::PexprExpand, but this will produce != operators in an
array NOT IN case which cannot be interpreted by the DXL translator. The
choice of using a ConstraintInterval::PexprConstructDisjunctionScalar
instead is that it will only produce greater than or less than or equal
to comparisons which can be handled.

(2) Do not expand array expressions in a PredicateUtils helper function which
was ran as part of an expression tree pre-DXL translation cleaning. Now
the change in (1) will handle this case. This removes redundancy in the
resulting plan.

A traceflag has been added to enable the PredicateUtils change. A test has
been modified to show the effect of enabling the traceflag.

When the traceflag is eventually removed, there will be several more tests
that need updating. This will come in a later PR.