[#125960675] Preprocessing stage for Array Conversion
Adds an expression preprocessing stage to convert applicable expressions
into array IN or NOT IN expressions. This creates more succinct
expression trees and reduces redundancy in final expression output.

This feature works by doing a recursive walk of the expression tree  which
looks for multiple AND or OR children which are either array expressions or
ScalarComparison children. It then converts them into Constraints.
Constraints hold logic to detect redundancy and also to convert OR/AND
statements into array statements. We chose to put this method after step
8 of expression preprocessing because step 9 will propagate constraints,
therefore it makes sense to propagate array constraints instead of OR
constraints because they are more succinct and also preprocessing will
attempt to re-derive constraints from propagated constraints which can
lead to redudant expressions being created if array constraints are not
passed.

This feature is controlled by the EnableArrayDerive traceflag.