Remove last remnants of gp_enable_delete_as_truncate GUC.
The feature was disabled, by removing it from guc.c, only months after
it was introduced, back in 2009. I'm not sure if it ever made it into a
public release. We have no plans to resurrect it, and it has surely
bitrotted so badly that even if we did, it would be better to start from
scratch.