Fix annoyances in the regression test added for memtuple-forming bug.
Commit 785c9bea5524a4293e78d7b75bd5820bf23a4ac0, "Fix memtuple forming bug
when null-saved size is larger than 32763 bytes", added a regression test
that was annoying for multiple reasons:

* The 'aggregates' regression test is from the upstream. We're trying to keep
upstream test files unchanged, to make diffing and merging with the upstream
easier.

* The query is huge, about 250000 bytes, and it's all on one line. Not all
tools play well with such long lines. Emacs became really sluggish and the
laptop fan started screaming, when I moved the cursor over that line. When I
tried to copy-paste that into psql, my terminal ground to a halt. It took
minutes to process it, ie. just to paste the line to the terminal. (Note to
self: get a better terminal program :-) ).

* The query takes a very long time to run, about 30 seconds on my laptop. I
did some profiling: all the time is spent in the code in ExecInitAgg() that
checks for duplicate aggregates. That loop is O(n^2), and calls equal() on
the AggRef expressions on each iteration.

* When the query doesn't crash because of the bug that was fixed, it produces
an "invalid timestamp" error. Seems surprising that the expected result is
an error, unless you're specifically testing for error handling. It also
makes me worried that the test might fail to test the right thing in the
future, if the planner decides to rearrange things so that the cast to
timestamp is evaluated first, so that it errors out before even constructing
the array with the aggregate results.

To fix those issues:

* Move the test query to 'bfv_aggregate' test. BFV stands for Bug Fix
Verification, i.e. checking that old bugs haven't reappeared, and that's
exactly what this is. (There are other GPDB additions in 'aggregates' test
besides this, but let's not make it worse.)

* Split the test query into reasonably-sized lines.

* Replace the aggregated expression with a simpler one. I'm hesitant to try
optimizing the O(n^2) code, because code comes from the upstream. There
have been some changes in that area in recent versions of PostgreSQL, and
I'm not sure if the recent changes will make this better or worse, but not
much point spending time on it before we catch up. And then we should get
any improvements there committed to the upstream first.

Replacing the CASE-WHEN construct with something simpler doesn't fix the
O(n^2) behavior, but it makes the constant factor smaller, by making the
equal() calls cheaper. I replaced it with "c4 % [number]", and now the
query takes about 17 seconds on my laptop (was 30 s). That's still longer
than I'd wish, but every little helps.

* Remove the cast to timestamp that was causing the error on succes.

In the passing, also add a comment explaining what the purpose of the test
is, and some notes on the O(n^2) behavior.