Fix of PartitionSelector INNER VarNo reference by passing parent plan pointer.
In this PR we revert the earlier fix in #840, but keep some part of it. This
PR is similar in spirit to the earlier as it depends on parent of
PartitionSelector (which is expected to be Sequence) to deparse printablePredicate.
However, this PR resolves the parent from child, instead of resolving child from
parent and localizes the changes to just explain.c. We also keep the old behavior
of clearly isolating Sequence node from PartitionSelector node.

Besides this, we also pass the parent to resolve both outer and inner varno of a
PartitionSelector's printablePredicate, which is also necessary to make bfv_partition green.