Remove dead function.
This is remnants of an old configuration file format that pre-dates the
RSA BSAFE lockbox implementation. I'm actually about to replace the current
lockbox implementation with something that resembles this dead function, but
it's not going to be quite the same, so let's get this dead function out of
the way first.