Clean up the compiler and linker options used to build ddboost stuff.
Passing CPPFLAGS when building libgpbsa75.so and libgpbsa71.so allows the
compilation to work, if ddboost header files are in a directory specified
in "./configure --with-includes=...". Using CFLAGS_SL rather than plain
CFLAGS ensures we use the correct flags determined by autoconf, for building
a shared library. That includes -fPIC on platforms that need it, so we don't
need to pass that explicitly anymore.

Move $(LIBS) after all the other libraries in all the command lines.
Doesn't make a difference right now, but with the upcoming patch to get rid
of the dependency to RSA BSAFE, it seems to be necessary that "-ldl" is
passed after -lDDBoost. Apparently, DDBoost.so uses dlopen() and friends,
and gets upset unless -ldl is passed after -lDDBoost (-ldl is part of LIBS,
at least on my system). This is probably highly platform-dependent, but I
think this order should work everywhere, even if it doesn't matter for
some platforms.