Fix compiler warning.
The compiler rightly pointed out that this memset() call was not zeroing
the whole struct, as it was clearly intended to. 'dd_options' is a pointer,
so sizeof(dd_options) is only 4 or 8, depending on architecture. That was
harmless, because the caller just zeroed out the array, but let's silence
the compiler warning. Zeroing it twice is not necessary, but this is not
at all performance-critical, so I don't want to second-guess the intentions
of the original coder right now.