Fix persistent rebuild tool issue when mirror data dir is empty
During pt rebuild when the mirror is down and the mirror data directory has some missing files,
persistent rebuild will fail with the error "missing files from source".
To mitigate this, we will be skipping the pertsistent rebuild for mirror if it is down.
We don't need to backup the files if the segment is already down anyways.

Add unit tests and behave test.

Authors: Chumki Roy & Marbin Tan