Adds a GUC to enable array constraint derivation.
This fixes the naming in c85f858ec725ec14a6e7870749f3a03fb2597310

A new feature of ORCA is to more efficiently handle array constraints.
It includes a new preprocessing stage, and a new way of internally
representing array constraints. This feature can be enabled by use of
this GUC.