Enhance cross verifying EOF for AO/CO with PT for EOF=0 case as well
Currently, during insert to AO/CO tables check is performed to validate
file-offset (EOF) found from pg_aoseg which would be used to start writing
the data to file. The offset to start writing needs to be greater than
currently marked EOF in persistent table. The validation is currently
performed for all other scenarios except EOF=0, so this commit is modifing to
enable it for EOF-0 case as well. As it was seen in field for yet to be known
reasons maybe indexing issues on pg_aoseg table, that incorrect EOF=0 was
fetched from aoseg and valid data was over-written. So, this protection will
prevent data-loss scenarios.