Fixes a bug in CConstraintInterval involving NULLs
When a CConstraintInterval was constructed using an array containing
CDatumGenericGPDB types we exposed a bug where the comparison logic with NULLS
was different between datums which could be evaluated within ORCA.

The fix is straightforward: check for NULL datums in an array when
constructing a CConstraintInterval. Set a flag in the
CConstraintInterval class and do not add the NULL datum to the data
structures inside of the Interval.

Also adds a unit test.