Remove tidycat.
System catalogs are now defined like in the upstream, without any special
tidycat headers.

Note: This doesn't change the way pg_proc_gp.h is generated from
pg_proc.sql, by catullus.pl.

This comes with a replacement for generating 4.3.json:
src/backend/catalog/process_foreign_keys.pl. The foreign keys were the
only thing that the json file was used for. AFAIK gpcheckcat is the only
tool that reads that file, and it only paid attention to the foreign key
information.

While working on this, I noticed that a few tables were missing foreign
key declarations. I added FIXME comments on them. Also, a few tables used
to have "vector" references, e.g. pg_index.indclass was an oidvector, where
each element of the array points to pg_class.oid. AFAICS, those
declarations were actually not used for anything. I left those in place as
comments, in case we want to add support in gpcheckcat and the new
process_foreign_keys.pl tool for them, but for now they're just
documentation.

This also removes pablopcatso.pl. It read the json file (Ok, so there was
one more user besides gpcheckcat for it), which is now gone. We could
create a tool that reads the same information straight from the header
files, or from a live database, but it was just a developer aid, and I
don't think anyone's used it for quite a while, so I don't think we need a
replacement.