Remove unused socket code, and ioutils::SzMkTemp function.
ioutils::SzMkTemp called mktemp(), which is a security risk. That caused
a compiler warning. The only thing that used ioutils::SzMkTemp was the
test code for CSocket, but all the socket code was unused anyway, so just
remove the whole lot.

I was also getting a warning/error on strict aliasing:

/home/heikki/gpdb/gpos/server/src/unittest/gpos/net/CSocketTest.cpp:549:27: error: dereferencing type-punned pointer will break strict-aliasing rules [-Werror=strict-aliasing]
   GPOS_ASSERT((*(ULONG *) szRecv) == pinitUds->m_ulId);
                           ^
So that's another small reason to get rid of this.