Prevent inconsistent update to persistent info in relcache entry.
As part of creating a new relfile node, for example during reindex, the
persistent TID and the serial number in the relation descriptor used to be
updated with new values immediately.

This commit prevents such updates leaving the description unchanged.  Instead,
the relation descriptor will be udpated upon next cache invalidation.

The new ICG test is added to its own file and not in gp_create_table.sql
because reindex of pg_class index seems to affect planner, causing other tests
to fail with "insufficient memory reserved for statement" error.