[#128239337] Ensure constraints are constrainable types in preprocessor
We ran into cases where the preprocessor step 'PexprConvert2In' would
try to create a constraint but fail and get a null value returned to it,
which would cause ORCA to crash. This was because the constraints
framework can only create constraints on certain types. By adding an
additional check for constrainable types we guarantee that a constraint
will be returned.

Also added an ASSERT to check for this condition.