[#128188689] Adds ArrayConstraints flag and fixes to several tests
As part of our effort to find bugs in the Array constraints feature, we
turned on Array Constraints by default and ran the ctest suite. The
tests which failed, we examined the changed DXL and reasons for failure
and determined that the changes to the plans were logically equivalent
to the original plan, simply modified to use arrays in some cases.
Turning the ArrayConstraints flag on for these tests will also exercise
the ArrayConstraints change, which is desirable.

We intentionally did not enable the flag for CArrayExpansionTest. It
requires the flag to be off to test its intended code path.