Do not coalesce if count is not generated by quantified subqueries
Given DDL
```
CREATE TABLE foo (a, b) AS SELECT i,i FROM generate_series(1,1) AS t(i) DISTRIBUTED BY(a);
CREATE TABLE bar (c int, d int);
```
And a query
```
SELECT (SELECT COUNT(*) FROM bar GROUP BY c LIMIT 1) FROM foo;
```

ORCA used to generate a plan that coalesces the result of `count` to
zero. This was wrong. The wrong result looks like this:

```
 ?column?
----------
        0
(1 row)
```

And the portion of the code that adds the `coalesce`
was difficult to understand and we found no test case to justify it.

The expected result should be like this:
```
 ?column?
----------

(1 row)
```

This commit replaces the `coalesce` with a identity projection. This is
still ugly because we are generating a superfluous `project` operator,
but it is at least correct.

We may follow up with a clean up that eliminates this superfluous
projection.
[#127333267]