Replace CaQL calls with plain syscache calls in executor.
We surely want to have a syscache for any lookups done at execution, for
speed. This makes it more clear that there indeed is a syscache for these.
(And I'm trying to remove CaQL altogether, anyway)

The CaQL lookup in assign_func_result_transient_type(Oid) was in fact
wrong before. It called caql_endscan(), and continued to access the tuple
after that. Fix that while we're at it.