Refactor IndexInfo struct population code.
The IndexInfo structs are used by ORCA to decide which indexes match an
expression, and there may be some other similar use for them too. This patch
changes it to fetch the index expressions and predicates from the relcache,
rather than deparsing them directly from the pg_index catalog. The versions
stored in the relcache have been run through const-evaluation and
canonicalization passes, so they are slightly different. I believe the
versions stored in the relcache are what we actually want here, for the
reasons explained in RelationGetIndexPredicate() and
RelationGetIndexExpressions().