Remove UDP signalling code intended to work around OS X bug
The UPD signalling code was added due to a bug in the pthreads code
in OS X 10.6 Snow Leopard which prevent pthread_cond_timedwait() to
work as intended. This version of the OS was discontinued in 2011
so it's time to retire this work around. OS X is not a supported
GPDB platform and as such we only make a best effort to run it on
the currently latest version.

In the process also make OS X use _timedwait() and not the relative
function for simplicity.