Remove gp_flush_buffer_pages_when_evicted GUC.
Looking at old git history, this was added back in 2009. The related
ticket on adding it said:

    Add GUC that make every buffer page from the buffer pool flushed on
    eviction.

    Note that this will NOT necessarily flush all buffer pages when the
    postmaster is shutdown. I think this is acceptable for our purposes.

    (Our purpose is to make sure that overwrites of the buffer pages are not
    lost and instead are always written to disk so we can catch errors)

I'm not sure what errors that was meant to catch, or how, but I don't
think we have any regression tests or anything that uses it anymore.
Let's remove it, to make merging with upstream easier.