Reformat some GPDB-specific files.
Whitespace and comment fixes, to follow the usual project style.

Remove duplicated function comments between the .h and .c files. Per the
usual project convention, all explanations of a function and its arguments
are in the .c file, and the .h file only contains the prototypes. There
were some additional comments about the "sections" of the files that
seemed useful, but were only in the .h files. I moved those to the .c
files instead.