Avoid dereferencing zero in the faultinjector
Apply the same fix as in 48983e3 which missed these invocations of
zero derefs. Dereferencing address zero is undefined behavior in C
and the compiler is free to optimize any such calls away (even though
that is believed to be rare). Ensure that any optimizations are
avoided by declaring the variable volatile and at the same time get
rid of a pesky clang compiler warning.