Only distribute on hashable columns [#129098219]
Previously, ORCA may generate a (parallel append) plan that
redistributes on (among other columns) a column of an unhashable data
type. This plan is not executable by the executor. This commit fixes
that by adding a check on the hashability of the data type of each
column before generating the distribution spec for the motion.