Remove index usage for pg_aoseg and pg_aocsseg lookups.
AO and AOCO tables do an index lookup during inserts and compaction
against the segno column. Select statements do not use this
index. This patch will make it so that we now do a heap scan on the
pg_aoseg/pg_aocsseg tables to obtain the segment file information to
remove the work on maintaining the index. This should not impact
performance as there are only 128 tuples in the pg_aoseg/pg_aocsseg
tables and relieves some possible index corruption issues that may
uncommonly appear in rare cases.

Authors: Jimmy Yih, Asim Praveen, and Abhijit Subramanya