de-genericized and split `CHashedDistributions` [#130220363]
This should have been part of commit 01e9695 .
The class `CHashedDistributions` was extracted because we wanted to
re-use code from `CPhysicalSerialUnionAll` in `CPhysicalParallelUnionAll`.

Now that they have started to diverge, since commit 01e9695, there's no
more reason to share this class. In fact, having mixed usage of this
class and mixed logic proves to be riskier and more prone to errors.