Fix buffer overflow in AO table compression ratio calculation
When calculating the compression ratio a char[8] buffer was used for
intermediate storage via snprintf() before passing into float8in. For
large values on compression ratios this would trivially overflow and
crash the process. Fix by removing the char conversion altogether and
just perform the rounding in place. Also remove a memory context switch
that makes no sense in the current version (and made little sense in
the previous). For the AO col tables, move to using the scanint8()
function without going via the official int8in functionality to save a
few levels on the stack.

Below are the steps to reproduce that originally triggered the bug but
many variations (and simplifications) on this is possible:

    create table foo
	(
		a integer encoding
		(
			compresstype = rle_type,
			compresslevel = 1
		)
	)
	with (appendonly = true, orientation = column)
	distributed by (a);

	insert into foo select a from
	(
		select 1 a, generate_series(1,500000) b
	) s;
	insert into foo select a from
	(
		select 15 a, generate_series(1,500000) b
	) s;
	select get_ao_compression_ratio('foo');

While there fix a few small style nitpicks as well.