Changes to make compilation with gcc-6.2.0 successful
Orca translator changes reorder include statements so that system and
PostgreSQL headers are grouped at the top.  Including such headers
within a C++ namespace causes compilation to fail with the new GCC.

The gpmonlib change removes inline keyword.  The new compiler is not
happy with inlining a function defined in an external object.