Rename <,>,<= and >= operators for complex datatype
Due to the mathematical properties of complex numbers, the comparison
operators should be used with a great deal of caution and intent. To
avoid unintentional misuse, rename these operators prefixed with a
leading '<' or '>' depending on the operator. The choice of prefix is
to make these operators closer to the 2d point datatype operators.