Hint to disable CXformExpandNAryJoinDP when join has large number of children [#121181555]
CXformExpandNAryJoinDP uses dynamic programming (DP). It is an expensive
transformation when the number of join children is more than 10 since
computes all the subsets and join orders for
large number of tables.

In this commit we added a hint that sets a hard-limit on how many joins
we want to consider in the CXformExpandNAryJoinDP rule.

-- Venky and Ivan