Adding length checks for netbackup parameters in python and c code (#1159)
We have a limit of 127 characters for parameters passed to netbackup (--netbackup-service-host, --netbackup-policy, --netbackup-schedule, and --netbackup-keyword). We now allocate more memory for these parameters and have added code to check their length and throw errors if they are too long. These checks exist in both the python and c code for extra coverage. We also have unit and behave tests to ensure the new checks are working properly.