Correctly compare hashed distribution specs for equality [#130664383]
The equality function was written at a time when
`CDistributionSpecHashed` was at a leaf of the type hierarchy. That's
why it never bothered to check for the type. This was no longer true
after commits 6c7c5d5 and 0c11dcd where we introduced two additional
distribution specs that shallowly inherit from
`CDistributionSpecHashed` .

The lack of type check on both sides result in bugs when the motions
enforcing one distribution do not make it into the group because of the
presence of a **very** similar motion, e.g.

Given
```
CREATE TABLE foo (a int, b int) DISTRIBUTED BY (a);
CREATE TABLE bar (c int, d int) DISTRIBUTED BY (c, d);
```

We want to generate two alternatives for
```
SELECT a,b FROM foo UNION ALL SELECT c,d FROM bar;
```

0. The first one tries to mirror the output distributions:
  ```
                                             QUERY PLAN
  ------------------------------------------------------------------------------------------------
   Gather Motion 3:1  (slice3; segments: 3)  (cost=0.00..862.00 rows=1 width=8)
     ->  Append  (cost=0.00..862.00 rows=1 width=8)
           ->  Redistribute Motion 3:3  (slice1; segments: 3)  (cost=0.00..431.00 rows=1 width=8)
                 Hash Key: foo.a, foo.b
                 ->  Table Scan on foo  (cost=0.00..431.00 rows=1 width=8)
           ->  Redistribute Motion 3:3  (slice2; segments: 3)  (cost=0.00..431.00 rows=1 width=8)
                 Hash Key: bar.c, bar.d
                 ->  Table Scan on bar  (cost=0.00..431.00 rows=1 width=8)
  ```

0. The second one tries to generate motions that mirror the input
distributions:
  ```
                                             QUERY PLAN
  ------------------------------------------------------------------------------------------------
   Gather Motion 3:1  (slice3; segments: 3)  (cost=0.00..862.00 rows=1 width=8)
     ->  Append  (cost=0.00..862.00 rows=1 width=8)
           ->  Redistribute Motion 3:3  (slice1; segments: 3)  (cost=0.00..431.00 rows=1 width=8)
                 Hash Key: foo.a
                 ->  Table Scan on foo  (cost=0.00..431.00 rows=1 width=8)
           ->  Redistribute Motion 3:3  (slice2; segments: 3)  (cost=0.00..431.00 rows=1 width=8)
                 Hash Key: bar.c, bar.d
                 ->  Table Scan on bar  (cost=0.00..431.00 rows=1 width=8)
  ```

But we will fail to generate the first plan because we would mistake one
motion ("no-op hash Redistribute on c,d") for the other ("strict hash
Redistribute on c,d").

This commit fixes that.

This potentially fixes [#130120913]