Use slot_getattr wrapper function for regular version.(#1194)
Slot_getattr in header file cause to have different pointer.
This cause issue when we register as external function. LLVM
expect the name of the external function to be unique for
each pointer. In order to fix it, we introduced wrapper function
in codegen_wrapper.cc and use that instead of slot_getattr.