Remove scalar subquery with outer ref and single const tbl child[#131019665]
This commit introduces a preprocessing feature where there is a scalar
subquery with a single constant table child. The subquery must have an
outer ref different than the output column of the CLogicalConstTableGet.

Instead of having the subquery we replace it with a CScalarIdent

Below is the input and output of the sample.

+--CScalarSubquery["a" (0)]
      |--CLogicalConstTableGet Columns: ["" (18)] Values: [(1)]
+--CScalarIdent "a" (0)