Cost model favors parallel union all [#130121327]
We enable both `serial` and `parallel union all` in the search space
and we favor `parallel union all` if the `optimizer_parallel_union`
GUC is enabled. `parallel union all` is costed as the same as maximum
costed child instead of adding the children's cost all together.

This makes the framework to choose the parallel union all instead of
serial union all when both operators are part of valid plan
alternatives.