Enable both parallel and serial union [#130121327]
Previously, when `optimizer_parallel_union` GUC is set, we were only
generating plans with `parallel union all` and omitting `serial union
all`. In this commit, we generate plans for both `serial union all` and
`parallel union all`.

In some cases, enforcement framework may not let `parallel union all` to
be part of a plan because of newly introduced distribution specs and
motions it generates. It is better to provide the legacy union all
implementation to be part of the alternatives instead of completely
omitting the alternative.

Signed-off-by: Xin Zhang <xzhang@pivotal.io>