Make return value of GetServerProcessTitle() const and fix style
The returned string is better declared as a const char pointer, update
and fix the two call sites. Also add the walwriter process to the list
and make the if/else chain consistent by just using if. Any if statement
being true will leave the function anyways so might as well make it
easier to read.