Style, spelling and formatting cleanup of postmaster
Wholesale cleanup of the postmaster.c file as it was worked on. Some
of the fixups include: function return value declarations on their own
line; breaking long lines; concatenating broken lines that collapsed
are less than 79 chars; fixing spelling; moving MPP-XXX markers to end
of the comments rather then prefixing the comment; ..and more.

Since the file originates from upstream, opt for the postgres style
when in conflict to keep the file consistent.