Make status report feedback not depend on debug printing
The tracking of process startups was in a past commit encompassed
within an if brace block seemingly by accident. This seems to have
been copy-pasted with the result being that tracking processes was
depending on the debug_print_server_processes GUC being set. This
separates the tracking and the process checking and also removes
a few checks that weren't necessary.