Fix UNION query failed to generate plan when parallel union is enabled [#132718373]
Orca failed to generate plans for the following query when
parallel union is enabled:

SELECT * FROM foo UNION SELECT * FROM bar;

Because returning EpetRequired in the EpetDistribution of CPhysicalParallelUnionAll
causes the enforcement framework to falsely introduce motions that can
cause failure in orca to generate plans for some queries.

When we introduce unnecessary CPhysicalMotionHashDistribute on top of
CPhysicalParallelUnionAll, the system actually detects that it is unnecessary.
However, the optimization flow gets corrupted because of that since
COptimizationContexts are indeed related to the each other. One optimization
context can be a child or a parent of another optimization context in another group.

To resolve the issue, share the same logic with the CPhysicalSerialUnionAll
by lifting the CPhysicalSerialUnionAll::EpetRequired code to the parent
class CPhysicalUnionAll.

Closes #116