Remove "-funroll-loops finline-limit=1800" from standard CFLAGS.
Aggressive inlining can improve performance in hot spots, but when applied
indiscriminately, it makes binaries larger for little gain. For lack of
better evidence, let's trust the compiler defaults. This makes the build
faster, and the binaries smaller. If we find out that these options are
beneficial for some specific functions or source files, let's add them
back to those special cases.

In a quick test on my laptop, this makes the build about 2x faster, and
shrinks the binaries from about 14MB to 9MB (after running "strip").