Fix wrong virtual tuple check in codegened slot_getattr
The code that we generate for slot_getattr is not correct.
In particular, to check if a tuple is virtual, we have to implement the code below:
if (TupHasVirtualTuple(slot) && slot->PRIVATE_tts_nvalid >= attnum)

In the codegened slot_getattr, we had not implemented the
second condition, i.e., slot->PRIVATE_tts_nvalid >= attnum.

In this commit, we generate code for the second condition.