Protection for transactional stat updates only if in transaction
Transactional stats for heap_delete must be updated only if in transaction,
which should always be the case. Except the issue was encountered now when we
started calling heap_delete for PT tables to free tuples instead of older
mechanism. During recovery based on object state if transaction was aborted the
tuple in PT needs to be deleted and was failing in this function as
TopTransactionContext is not allocated. Hence, added the protection that only if
nesting level if greater than 0 which will be the case if we are in transaction
collect stats else ignore the same. Which helps to fix the problem.