Do not ignore NULLs when computing a DISTINCT array_agg.
Currently GPDB ignores NULL values when computing select array_agg(DISTINCT a).
This doesn't comply with the SQL-2008 standards (Refer
https://www.postgresql.org/message-id/e08cc0400911150005r82bc48du726d02b680c7920c%40mail.gmail.com
for context).
This was fixed in postgres 9.0 in 34d26872ed816b299eef2fa4240d55316697f42d.
This commit brings GPDB output to match that of upstream postgres.

This fixes issue #427.