Some functions no longer need to return newly-created object's OID.
Now that we capture and dispatch OIDs separately, we can revert these
changes vs. upstream.

I wasn't 100% sure if the change in bmbulkdelete() was OK. We no longer
update the relNode in the relcache entry. However, running this in a
debugger shows that reindex_index() updates the relcache entry itself,
so setting it in bmbulkdelete should not be necessary. Also, the
regression tests are happy with this.