Making expression tree mutation unconditional (#1273)
* Making expression tree mutation unconditional

We previously did not consider that we need to mutate an expression tree
only if there is a pseudo column. This introduces an executor crash in
non-assert build as executor does not know how to interpret a pseudo
column. Moreover, in assert build we fail an assert.

The conditional we are removing  was introduced to optimize performance
by commit a36436ea3f402293bc5484b83823eed8d6a8dfc3, but a quick performance
run shows that performance is not noticeably affected by removing the
optimization. [#131785597]

Signed-off-by: Foyzur Rahman <foyzur@gmail.com>

* Adding PR comments.