Remove ORCA-specific expected output.
The test is failing with ORCA, because the expected output file hasn't been
updated for the spelling of "at OPEN". The regular expected output file was
updated in commit 05590cbf, but the ORCA-specific one was not.

There is no meaningful difference between the ORCA and non-ORCA outputs,
so just remove the alternative output file.