Insert random motion for randomly distributed children for parallel append [#133978027]
Parallel Union currently has 2 optimization request. The second request is not satisfied
when the children of the parallel union are randomly distributed.

For the first request, if the output columns are not hashable columns the
`CPhysicalParallelUnion` requires random distribution. Since the children are randomly
distributed the `CPhysicalTableScan` already satisfied the optimization request and no
motion is created on top of `CPhysicalTableScan`.

To overcome this issue and enforce a `CPhysicalMotionRandom`, we introduce a new distribution
spec called `CDistributionSpecRandomStrict`. This distribution spec is requested replaces
regular `CDistributionSpecRandom` in the first optimization request where no output column is
redistributable.

Below is the output after this commit:
```
explain select xmin from foo union all select xmin from bar;

Physical plan:
+--CPhysicalMotionGather(master)   rows:1   width:4  rebinds:1   cost:431.000029   origin: [Grp:2, GrpExpr:3]
   +--CPhysicalParallelUnionAll   rows:1   width:4  rebinds:1   cost:431.000014   origin: [Grp:2, GrpExpr:1]
      |--CPhysicalMotionRandom   rows:1   width:38  rebinds:1   cost:431.000014   origin: [Grp:0, GrpExpr:2]
      |  +--CPhysicalTableScan "foo" ("foo")   rows:1   width:38  rebinds:1   cost:431.000007   origin: [Grp:0, GrpExpr:1]
      +--CPhysicalMotionRandom   rows:1   width:38  rebinds:1   cost:431.000014   origin: [Grp:1, GrpExpr:2]
         +--CPhysicalTableScan "bar" ("bar")   rows:1   width:38  rebinds:1   cost:431.000007   origin: [Grp:1, GrpExpr:1]

                                           QUERY PLAN
------------------------------------------------------------------------------------------------
 Gather Motion 3:1  (slice3; segments: 3)  (cost=0.00..431.00 rows=1 width=4)
   ->  Append  (cost=0.00..431.00 rows=1 width=4)
         ->  Redistribute Motion 3:3  (slice1; segments: 3)  (cost=0.00..431.00 rows=1 width=4)
               ->  Table Scan on foo  (cost=0.00..431.00 rows=1 width=4)
         ->  Redistribute Motion 3:3  (slice2; segments: 3)  (cost=0.00..431.00 rows=1 width=4)
               ->  Table Scan on bar  (cost=0.00..431.00 rows=1 width=4)

```

Signed-off-by: Xin Zhang <xzhang@pivotal.io>