Use proper error code for NULL constraint violations in ORCA.
When a "null value in column <col> violates not-null constraint" error was
thrown from the ORCA translator code, it was propagated to the rest of the
system with ERRCODE_INTERNAL_ERROR. That's not nice, if an application
wants to e.g. catch errors based on the error code.

The reason to do this right now was that my previous commit that moved
some ORCA translator code in the source files made the regression tests
to fail. That's because an error handling code adds the source file name
and line number to the error message, like (CTranslatorUtils.cpp:2627),
and we had some such line numbers memorized in the regression tests.
When the code moved, the tests broke. That's obviously too fragile, and
using a proper error code fixes that by preventing the line number from
being printed. That also removed the last difference between the ORCA
and non-ORCA output from the 'insert' regression test.

We could probably do the same for other errors thrown from ORCA, but I'll
leave that for another day.