Move some tests for append-only tables out of bugbuster.
The 'ao_create_alter_valid_table' test in bugbuster was completely defunct,
as the expected output just contained an error about connecting to a
non-existent database. However, the exact same test queries were included
in the schema_topology test. Schema_toplogy is huge, and some of the tests
in it require a separate output file for ORCA, which these test queries do
not, so splitting it up into smaller parts seems like a good idea.