Ensure that resources are freed before reusing pointers
On the first iteration we know the structs are empty, but when coming
into create_gang_retry we must ensure that they have been freed. Move
to Assertions rather than setting NULL since that's already performed
before returning to the retry label.

Initially per a false positive gripe in Coverity.