Remove silly unit tests on GUC.
Without assertions, the only thing that was checked was that
numProcsPerSegment() returns 1, when the
gp_resqueue_priority_cpucores_per_segment is 1. That seems excessive, as
numProcsPerSegment() is a trivial function that just returns the value
of gp_resqueue_priority_cpucores_per_segment.

With assertions, it checked that an assertion is thrown if
gp_resqueue_priority_cpucores_per_segment has an invalid value. That seems
excessive, because those assertions are just extra sanity checks. The GUC's
minimum and maximum values prevent it from having invalid values in the
first place.

Linking each mock test program takes a few seconds, so this shaves some
time off the build+unit test process.