Allow incremental backup for ddboost to work across multiple dates. (#1328)
When running backup with ddboost, we now use dump timestamp instead of current timestamp as the parent directory.  This fixes a bug where the dump fails if the report and cdatabase files are generated on different days (e.g. if the dump is kicked off just before midnight).

Authors: Stephen Wu and Jamie McAtamney