Move catalog test out of parallel group (#1338)
pg_get_constraintdef() looks up information on the constraint
from the syscache, which uses SnapshotNow to fetch the information. But the
query itself uses an MVCC snapshot. If the query's MVCC snapshot sees a
constraint that is concurrently dropped so that it's not visible to SnapshotNow,
pg_get_constraintdef() will report error.

Thanks Heikki for above description.