Update comment to explain why we need to insert Result nodes here.
I figures out the reason this is needed. It was alluded to by the
"This is so the executor can use a faster path to evaluate the  given
node's targetlist" comment, but hopefully this new comment makes it
more clear what the "faster path" is.

Remove the GPDB_83_MERGE_FIXME comment. There's a new mechanism for
dealing with SRFs in targetlists in the works for PostgreSQL 10
as of this writing, but until then, I think we'll live with this.